# openapi.yaml
openapi: 3.0.3
info:
  title: TiffinBook API
  description: >
    Subscription management and pro-rated billing for home-style tiffin
    (lunch delivery) services. All `/api/customers/*` routes are scoped to
    the logged-in owner and require a Bearer token from register/login.
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
paths:
  /api/auth/register:
    post:
      summary: Create an owner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: "owner@example.com" }
                password: { type: string, example: "hunter22" }
      responses:
        '201': { description: 'Account created; returns token and email' }
        '400': { description: Invalid input, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Email already registered }
  /api/auth/login:
    post:
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: 'Returns token and email' }
        '401': { description: Invalid credentials }
...